backend/foodgram_project/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2.13
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.4
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
pep8-naming==0.12.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
